requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
segment-anything>=1.0
transformers>=4.35.0
supervision>=0.16.0
scikit-learn
roboflow